Change order of steps
`Start your server 'npm start' and check if your bootstrap import works as well as the test button! If all is displaying then we are almost done. We are sooooo close!` - button is not displayed since the `Button` component isn't imported and used in `App.js` yet. 

It feels more natural to first have button implemented and then work on the API call.